Move previously inner classes to outer scope
