+++ NFluent.35/Extensions/ExtensionsCommonHelpers.cs
@@ -33,7 +33,12 @@ namespace NFluent.Extensions
         /// <param name="theObject">The theObject.</param>
         /// <returns>A string that represents the current object. If the object is already a string, this method will surround it with brackets.</returns>
         public static string ToStringProperlyFormated(this object theObject)
-        {           
+        {
+            if (theObject == null)
+            {
+                return "null";
+            }
+
             if (theObject is char)
             {
                 return string.Format("'{0}'", theObject);
