+++ NFluent.35/NFluent.35.csproj
@@ -63,6 +63,9 @@
     <Compile Include="Assertions\DoubleSpecificCheckExtensions.cs" />
     <Compile Include="Assertions\EventWaitHandleCheckExtensions.cs" />
     <Compile Include="Assertions\ObjectFieldsCheckExtensions.cs" />
+    <Compile Include="Extensibility\InstanceBlock.cs" />
+    <Compile Include="Extensibility\IValueBlock.cs" />
+    <Compile Include="Extensibility\ValueBlock.cs" />
     <Compile Include="FluentCodeCheck.cs" />
     <Compile Include="Extensibility\ExtensibilityHelper.cs" />
     <Compile Include="Assertions\IntSignedCheckExtensions.cs" />
