+++ NFluent.40/NFluent.40.csproj
@@ -149,6 +149,15 @@
     <Compile Include="..\NFluent.35\Extensibility\IChecker.cs">
       <Link>Extensibility\IChecker.cs</Link>
     </Compile>
+    <Compile Include="..\NFluent.35\Extensibility\InstanceBlock.cs">
+      <Link>Extensibility\InstanceBlock.cs</Link>
+    </Compile>
+    <Compile Include="..\NFluent.35\Extensibility\IValueBlock.cs">
+      <Link>Extensibility\IValueBlock.cs</Link>
+    </Compile>
+    <Compile Include="..\NFluent.35\Extensibility\ValueBlock.cs">
+      <Link>Extensibility\ValueBlock.cs</Link>
+    </Compile>
     <Compile Include="..\NFluent.35\Extensions\EnumerableExtensions.cs">
       <Link>Extensions\EnumerableExtensions.cs</Link>
     </Compile>
@@ -354,7 +363,7 @@
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <ProjectExtensions>
     <VisualStudio>
-      <UserProperties ProjectLinkerExcludeFilter="\\?desktop(\\.*)?$;\\?silverlight(\\.*)?$;\.desktop;\.silverlight;\.xaml;^service references(\\.*)?$;\.clientconfig;^web references(\\.*)?$" ProjectLinkReference="3a474510-f384-41ce-9270-539b1e51da7e" />
+      <UserProperties ProjectLinkReference="3a474510-f384-41ce-9270-539b1e51da7e" ProjectLinkerExcludeFilter="\\?desktop(\\.*)?$;\\?silverlight(\\.*)?$;\.desktop;\.silverlight;\.xaml;^service references(\\.*)?$;\.clientconfig;^web references(\\.*)?$" />
     </VisualStudio>
   </ProjectExtensions>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
